[#63] Add buildSrc
