+++ buildSrc/src/main/java/FileExt.kt
@@ -1,0 +1,12 @@
+import java.io.File
+import java.util.*
+
+fun File.loadGradleProperties(fileName: String): Properties {
+    val properties = Properties()
+    val signingProperties = File(this, fileName)
+
+    if (signingProperties.isFile) {
+        properties.load(signingProperties.inputStream())
+    }
+    return properties
+}
